+++ src/utils/FixedSchemaEndpointResolver.js
@@ -1,4 +1,5 @@
 const memoize = require('memoizee');
+const { matchPath } = require('./internal/pathMatcher');
 
 class FixedSchemaEndpointResolver {
     constructor(fixedRoute) {
@@ -12,32 +13,12 @@ class FixedSchemaEndpointResolver {
 
 function getMethodSchemaInternal(schemas, fixedRoute, method) {
     const methodLowerCase = method.toLowerCase();
-    const routePath = pathMatcher(schemas, fixedRoute);
+    const routePath = matchPath(schemas, fixedRoute);
     const route = schemas[routePath];
 
-    if (route && route[methodLowerCase] && route[methodLowerCase]) {
+    if (route && route[methodLowerCase]) {
         return route[methodLowerCase];
     }
-}
-
-function pathMatcher(routes, path) {
-    return Object
-        .keys(routes)
-        .find((route) => {
-            const routeArr = route.split('/');
-            const pathArr = path.split('/');
-
-            if (routeArr.length !== pathArr.length) return false;
-
-            return routeArr.every((seg, idx) => {
-                if (seg === pathArr[idx]) return true;
-
-                // if current path segment is param
-                if (seg.startsWith(':') && pathArr[idx]) return true;
-
-                return false;
-            });
-        });
 }
 
 module.exports = FixedSchemaEndpointResolver;
