+++ src/utils/SchemaEndpointResolver.js
@@ -1,4 +1,5 @@
 const memoize = require('memoizee');
+const { matchPath } = require('./internal/pathMatcher');
 
 // This logic is wrapped into class to have isolated memoization contexts
 class SchemaEndpointResolver {
@@ -9,32 +10,12 @@ class SchemaEndpointResolver {
 
 function getMethodSchemaInternal(schemas, path, method) {
     const methodLowerCase = method.toLowerCase();
-    const routePath = pathMatcher(schemas, path);
+    const routePath = matchPath(schemas, path);
     const route = schemas[routePath];
 
     if (route && route[methodLowerCase]) {
         return route[methodLowerCase];
     }
-}
-
-function pathMatcher(routes, path) {
-    return Object
-        .keys(routes)
-        .find((route) => {
-            const routeArr = route.split('/');
-            const pathArr = path.split('/');
-
-            if (routeArr.length !== pathArr.length) return false;
-
-            return routeArr.every((seg, idx) => {
-                if (seg === pathArr[idx]) return true;
-
-                // if current path segment is param
-                if (seg.startsWith(':') && pathArr[idx]) return true;
-
-                return false;
-            });
-        });
 }
 
 module.exports = SchemaEndpointResolver;
